Initial refactor.
